making a separate util function for resolving channel in session attributes
